requirements.txt
attrs==20.3.0
certifi==2020.12.5
grpcio==1.37.0
grpcio-tools==1.37.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-grpc==0.8.0
six==1.15.0
toml==0.10.2
grpc-interceptor==0.13.0
pydub==0.25.1
librosa==0.8.1
auditok==0.2.0
pathlib==1.0.1
youtube-dl
requests
responses
pytest-mock
pafy
pysrt
flask
Flask-Cors==1.10.3
tqdm
spectralcluster==0.1.0
git+https://github.com/Sujit27/speaker-identification
denoiser